pyproject.toml
[project]
name = "toolsmith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=2.6.1",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "serpapi>=0.1.5",
    "tqdm>=4.67.1",
]
